+++ man-roxygen/coerce.R
@@ -1,0 +1,2 @@
+#' @param coerce [\code{logical(1)}]\cr
+#'  If \code{TRUE}, the input \code{x} is returned as integer after an successful assertion.